pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = "A homemade chatGPT using OpenAI's API"
authors = ["jiajiabinx"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.3"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
azure-cosmos = "^4.3.1"
azure-identity = "^1.12.0"

[tool.poetry.dev-dependencies]
Flask = "^2.2.3"
openai = "^0.27.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
